requirements.txt
streamlit
PyMuPDF
sentence-transformers
chromadb
scikit-learn
Pillow
pdf2image
langchain
langchain-community
langchain-core
langchain-ollama
